Fix shared OpenStack provider subresources synchronization [SENTRY-2590]
Due to recent enforcement of cache key generation it is required that
either admin or tenant_id is specified when OpenStack client is constructed.
Although networks and subnets are related to particular tenants,
its context in pulling operations was not clearly specified as admin.
It lead to the failure in assertion check.